make separate button component
